+++ lib/spandx/parsers/sln.rb
@@ -2,26 +2,38 @@ module Spandx
   module Parsers
     class Sln < Base
       def parse(file_path)
-        paths = IO.readlines(file_path).map do |line|
-          next unless line.match?(/^\s*Project\(/)
-
-          path = line.split('"')[5]
-          next unless path
-
-          path = path.tr("\\", "/")
-          next unless path.match?(/\.[a-z]{2}proj$/)
-
-          dir = File.dirname(file_path)
-          path = File.join(dir, path)
-          Pathname.new(path).cleanpath.to_path
-        end.compact
-        paths.map do |path|
-          parser = Parsers.for(path, catalogue: catalogue)
-          parser.parse(path)
+        project_paths_from(file_path).map do |path|
+          Parsers
+            .for(path, catalogue: catalogue)
+            .parse(path)
         end.flatten
       end
 
       private
+
+      def project_paths_from(file_path)
+        IO.readlines(file_path).map do |line|
+          next unless project_line?(line)
+
+          path = project_path_from(line)
+          next unless path
+
+          path = File.join(File.dirname(file_path), path)
+          Pathname.new(path).cleanpath.to_path
+        end.compact
+      end
+
+      def project_line?(line)
+        line.match?(/^\s*Project\(/)
+      end
+
+      def project_path_from(line)
+        path = line.split('"')[5]
+        return unless path
+
+        path = path.tr("\\", "/")
+        path.match?(/\.[a-z]{2}proj$/) ? path : nil
+      end
     end
   end
 end